chore: separate add-ons from base node
